Rename `Sha1` to `Sha1Hash`
