+++ src/game/Anticheat/WardenAnticheat/WardenKeyGeneration.h
@@ -5,7 +5,7 @@
  */
 
 
-#include "Crypto/Hash/Sha1.h"
+#include "Crypto/Hash/Sha1Hash.h"
 
 #ifndef _WARDEN_KEY_GENERATION_H
 #define _WARDEN_KEY_GENERATION_H
