+++ src/shared/CMakeLists.txt
@@ -47,8 +47,8 @@ set (shared_SRCS
     Crypto/Hash/Hmac.cpp
     Crypto/Hash/HMACSHA1.h
     Crypto/Hash/HMACSHA1.cpp
-    Crypto/Hash/Sha1.h
-    Crypto/Hash/Sha1.cpp
+    Crypto/Hash/Sha1Hash.h
+    Crypto/Hash/Sha1Hash.cpp
     Database/Database.h
     Database/DatabaseCallback.h
     Database/DatabaseEnv.h
